requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
distro==1.9.0
Django==4.2.27
django-phonenumber-field==8.4.0
django-resized==1.0.3
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.7
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.14
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
oauthlib==3.3.1
openai==2.15.0
packaging==25.0
phonenumbers==9.0.21
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.2
python-decouple==3.8
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.8.3
sqlparse==0.5.5
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
